sia-backend/flask-server/requirements.txt
Flask==3.0.3;
uWSGI==2.0.27;
flask_favicon==0.1.4;
psycopg2-binary==2.9.10;
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
bcrypt==4.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
requests==2.31.0
SQLAlchemy==2.0.29
Werkzeug==3.0.2
WTForms==3.1.2
email-validator
Flask-Principal==0.4.0
Flask-User==1.0.2.2 #wirklich genau die Version
Flask-Mail==0.10.0
passlib>=1.7